Extracted default price scales to separate module to avoid circular dependencies
