+++ src/model/default-price-scale.ts
@@ -1,0 +1,8 @@
+export const enum DefaultPriceScaleId {
+	Left = 'left',
+	Right = 'right',
+}
+
+export function isDefaultPriceScale(priceScaleId: string): boolean {
+	return priceScaleId === DefaultPriceScaleId.Left || priceScaleId === DefaultPriceScaleId.Right;
+}
